adjusted parameters and chunks and overlap for better answers
