+++ src/eval/retrieve_bm25_eval_set.json
@@ -1,0 +1,128 @@
+[
+  {  "query": "hubble space telescope celebrated its 20th anniversary?",
+  "expected": [
+      "the hubble space telescope celebrated 20th anniversary April 24 2010."
+    ]
+  },
+  {
+    "query": "initially the hubble data transmissted to Earth how was it stored?",
+    "expected": [ "data was initially stored on the spacecraft when launched the storage facilities were old fashioned reel to reel tape drives but these were"
+    ]
+  },
+  {
+    "query": "what is spacex commonly referred as?",
+    "expected":["Space Exploration Technologies Corp", "space exploration technologies corp", "space exploration technologies corp commonly referred spacex is american space technology company headquartered starbase development site starbase texas since its founding 2002 company has made numerous advances rocket propulsion reusable launch vehicles human spaceflight satellite constellation technology of 2025 spacex is world s dominant space launch provider its launch cadence eclipsing all others including private competitors national programs like chinese space program spacex nasa united states armed forces work closely together means of governmental contracts spacex was founded elon musk 2002 vision of decreasing costs of space launches paving way sustainable colony mars 2008 falcon 1 successfully launched into orbit after three failed launch attempts company moved towards development of larger falcon 9 rocket dragon 1 capsule satisfy nasa s cots contracts deliveries international space station 2012 spacex finished all cots test flights began delivering commercial resupply services missions international space station also around that time spacex started developing hardware make falcon 9 first stage reusable company demonstrated first successful first stage landing"
+    ]
+  },
+   {
+    "query": "What are the legal implications of lunar resource extraction?",
+    "expected": [
+      "outer space treaty",
+      "non-appropriation principle",
+      "artemis accords",
+      "space resource utilization",
+      "international space law"
+    ],
+    "source_doc": "governing-the-moon-sp-2024-4559-ebook.pdf"
+  },
+  {
+    "query": "How do the Artemis Accords impact lunar governance?",
+    "expected": [
+      "artemis accords",
+      "governance",
+      "transparency",
+      "peaceful use",
+      "international cooperation"
+    ],
+    "source_doc": "governing-the-moon-sp-2024-4559-ebook.pdf"
+  },
+  {
+    "query": "What governance frameworks exist for managing activities on the Moon?",
+    "expected": [
+      "governance models",
+      "international treaties",
+      "lunar activities",
+      "responsible behavior",
+      "legal frameworks"
+    ],
+    "source_doc": "governing-the-moon-sp-2024-4559-ebook.pdf"
+  },
+  {
+    "query": "How can archaeology inform interstellar communication?",
+    "expected": [
+      "archaeological methods",
+      "decoding messages",
+      "material culture",
+      "cognitive inference",
+      "semiotics"
+    ],
+    "source_doc": "Archaeology_Anthropology_and_Interstellar_Communication_TAGGED.pdf"
+  },
+  {
+    "query": "Why is anthropology relevant to SETI?",
+    "expected": [
+      "cultural understanding",
+      "symbol systems",
+      "interpretive frameworks",
+      "cross-cultural communication",
+      "social context"
+    ],
+    "source_doc": "Archaeology_Anthropology_and_Interstellar_Communication_TAGGED.pdf"
+  },
+  {
+    "query": "What are the challenges of interpreting alien signals?",
+    "expected": [
+      "semantic uncertainty",
+      "cognitive bias",
+      "message encoding",
+      "semiotic barriers",
+      "lack of shared context"
+    ],
+    "source_doc": "Archaeology_Anthropology_and_Interstellar_Communication_TAGGED.pdf"
+  },
+  {
+    "query": "What was Korolev's role in the Soviet space program?",
+    "expected": [
+      "chief designer",
+      "Korolev",
+      "leadership",
+      "Soviet space program",
+      "missions"
+    ],
+    "source_doc": "rocketspeoplevolume4-ebook.pdf"
+  },
+  {
+    "query": "What was the N1 rocket and why did it fail?",
+    "expected": [
+      "n1 rocket",
+      "design issues",
+      "engine clustering",
+      "explosions",
+      "lunar competition"
+    ],
+    "source_doc": "rocketspeoplevolume4-ebook.pdf"
+  },
+  {
+    "query": "How did the USSR respond to the Apollo 11 mission?",
+    "expected": [
+      "apollo program",
+      "soviet reaction",
+      "lunar landing",
+      "propaganda",
+      "space race"
+    ],
+    "source_doc": "rocketspeoplevolume4-ebook.pdf"
+  },
+  {
+    "query": "Who owns the Moon?",
+    "expected": [
+      "outer space treaty",
+      "non-appropriation",
+      "no national ownership",
+      "international law",
+      "legal ambiguity"
+    ],
+    "source_doc": "governing-the-moon-sp-2024-4559-ebook.pdf"
+  }
+]
+
